conndlgs: look in /usr/lib/<cpu-os>
We have migrated most of out packages to debhelper 10, and now Debian
places our libraries in /usr/lib/<cpu-os>. We have to either search all
paths in LD_LIBRARY_PATH (appending conndlgs), decide to store our
conndlgs plugins directory in a lib path, or change the default path.

I've opted for the last solution so far, but I'm not super happy with
it. At least it makes it all work again.