iap-common: add default_custom_ui provider parsing
Service providers can be customised using the custom_ui path, but those
paths are always specific to the provider_id, which means that for every
provider id (which we see a configuration for a provider) would need the
icon names and such duplicated. This is messy, so instead introduce a
default_custom_ui path which is not service_id specific, but still
prefer the custom_ui path if it is set, allowing for per-config
customisation if the user so wishes.